Fix alignment OOM, migrate index to 64-bit, build bwa-mem2-style 2N index
Three foundational fixes that let bwa-rs index the full human genome and
map both strands, plus a ticket series for the remaining output-divergence
gaps vs bwa-mem2.

OOM (T-003): affine_extend_forward allocated a dense (query_len+1)x(ref_len+1)
DP matrix sized by the whole remaining reference even though the DP is banded,
so a single extension near a chromosome start allocated 100s of GB and was
OOM-killed. Cap the matrix width to the band (query_len + bandwidth); results
unchanged. chr1 alignment dropped from >128 GB (killed) to ~8.5 GB.

64-bit index (T-002): the SA/FM-index pipeline was 32-bit end to end
(libsais i32, u32 SA storage and on-disk format), capping references at
~2.147 Gbp so the human genome could not be indexed at all. Upgrade libsais-rs
0.1 -> 0.2 and use the 64-bit libsais64/libsais64_omp entry points; widen the
SA, f_column, occ counts, and the on-disk index format (VERSION 3 -> 4) to u64.
The mmap_index aux format stays 32-bit by design.

2N index + reverse strand (T-004): build the FM-index over forward ++
reverse-complement (Reference::as_slice_2n, FMIndex::build_2n, n_fwd), matching
bwa/bwa-mem2 so a single forward search covers both strands. Map SA positions
>= n_fwd back to forward coordinates with the 0x10 flag set and the CIGAR
reversed. On the chr1 verification set (300 unique pairs) the strand flag now
matches bwa-mem2 297/297 (was 0).

Also: IUPAC handling (T-001) and the remaining SAM-divergence gaps
(soft-clipping, QNAME truncation, paired-end, RNAME, SEQ/QUAL, CIGAR M,
tags, index persistence, multithreaded alignment, scoring/MAPQ/header) are
documented as T-005..T-016 in tickets.md.

All 231 tests pass.

+++ src/mmap_index.rs
@@ -131,7 +131,9 @@ impl MmapFMIndex {
         file.write_all(&bwt)?;
 
         for i in 0..len {
-            file.write_all(&(sa.get(i).unwrap_or(0)).to_le_bytes())?;
+            // mmap_index keeps a 32-bit on-disk SA format (≤4.29 Gbp); the main
+            // FMIndex path is the 64-bit one. SA values fit u32 for refs in range.
+            file.write_all(&(sa.get(i).unwrap_or(0) as u32).to_le_bytes())?;
         }
 
         // Build occ table from original sequence (not BWT) for correct counts
@@ -213,13 +215,13 @@ impl MmapFMIndex {
         (left, right)
     }
 
-    pub fn get_position(&self, sa_idx: usize) -> Option<u32> {
+    pub fn get_position(&self, sa_idx: usize) -> Option<u64> {
         if sa_idx >= self.len {
             return None;
         }
         let offset = self.sa_offset + sa_idx * 4;
         let pos = u32::from_le_bytes(self.mmap[offset..offset + 4].try_into().unwrap());
-        Some(pos)
+        Some(pos as u64)
     }
 
     pub fn count(&self, pattern: &[u8]) -> usize {
@@ -227,7 +229,7 @@ impl MmapFMIndex {
         right - left
     }
 
-    pub fn find_all(&self, pattern: &[u8]) -> Vec<u32> {
+    pub fn find_all(&self, pattern: &[u8]) -> Vec<u64> {
         let (left, right) = self.search(pattern);
         (left..right).filter_map(|i| self.get_position(i)).collect()
     }
